requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
gradio==4.16.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
pillow==10.2.0
openai==1.12.0